Fix some clippy warnings (#692)
* src/write/*: Fix clippy::redundant_guards
* src/read/aranges.rs: Fix clippy::try_err
* Fix clippy::unnecessary_cast

+++ src/read/value.rs
@@ -199,7 +199,7 @@ impl Value {
             Value::I32(value) => value as u64,
             Value::U32(value) => u64::from(value),
             Value::I64(value) => value as u64,
-            Value::U64(value) => value as u64,
+            Value::U64(value) => value,
             _ => return Err(Error::IntegralTypeRequired),
         };
         Ok(value)
